backend/requirements.txt
fastapi>=0.87.0
uvicorn>=0.19.0
SQLAlchemy>=1.4.44
python-dotenv>=0.21.0
passlib>=1.7.4
PyJWT>=1.7.1
fastapi-jwt-auth>=0.5.0
python-multipart>=0.0.5
bcrypt>=4.0.1
